Separating PaginatedEmbed from SwitchbladeEmbed
